Revert "Refactor(BE) : #43 피드백 반영한 수정"
This reverts commit 315f6d80abbff9b9b728ac52fef6398799474581.
